admin/acabado

+++ frontend/src/services/adminServices.js
@@ -1,11 +1,11 @@
 import api from './api';
 
 
-// GET http://localhost:3000/api/users (ALL)
-// GET http://localhost:3000/api/users/name (NAME)
-// POST http://localhost:3000/api/users (CREATE USER)
-// PUT http://localhost:3000/api/users/:id (EDIT USER)
-// DELETE http://localhost:3000/api/users/:id (DELETE USER)
+// GET http://localhost:3000/api/admin/users (ALL)
+// GET http://localhost:3000/api/admin/users/:user_id (ID)
+// POST http://localhost:3000/api/admin/users (CREATE USER)
+// PUT http://localhost:3000/api/admin/users/:user_id (EDIT USER)
+// DELETE http://localhost:3000/api/admin/users/:user_id (DELETE USER)
 
 // GET http://localhost:3000/api/mkt
 // GET http://localhost:3000/api/hr
@@ -13,32 +13,31 @@ import api from './api';
 //GET ALL USERS
 export const getAllUsers = async () => {
     try {
-        const response = await api.get(`/users`);
+        const response = await api.get(`admin/users`);
         return response.data;
     } catch (error) {
-        console.error('Error fetching all users:', error);
-        throw error;
+        throw error.response?.data || { msg: 'Error fetching All Users:'};
     }
 };
 
-//GET USERS BY NAME
-export const getUserByName = async(first_name)=>{
+//GET USERS BY ID
+export const getUserById = async(user_id )=>{
 
 
     try {
-        const response = await api.get(`/users/name/${first_name}`);
+        const response = await api.get(`admin/users/${user_id }`);
         return response.data;
     } catch (error) {
-        console.error('Error fetching user by name:', error);
-        throw error;
-    }  
+         throw error.response?.data || { msg: 'Error fetching user by ID:'};
+    }
+       
 }
 
 //POST USER -> Create
 
 export const createUser = async (userData) => {
     try {
-        const response = await api.post(`/users`,userData);
+        const response = await api.post(`admin/users`,userData);
         return response.data;
     } catch (error) {
         throw error.response?.data || { msg: 'Error creating new user' };
@@ -48,7 +47,7 @@ export const createUser = async (userData) => {
 //DELETE
 export const deleteUserById = async (user_id) => {
     try {
-        const response = await api.delete(`/users/${user_id}`);
+        const response = await api.delete(`admin/users/${user_id}`);
         return response.data;
     } catch (error) {
         throw error.response?.data || { msg: 'Error delete user' };
@@ -59,7 +58,7 @@ export const deleteUserById = async (user_id) => {
 
 export const updateUserById = async (user_id, userData) => {
     try {
-        const response = await api.put(`/users/${user_id}`, userData);
+        const response = await api.put(`admin/users/${user_id}`, userData);
         return response.data;
     } catch (error) {
         throw error.response?.data || { msg: 'Error edit user' };
